requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==4.2.17
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
fuzzywuzzy==0.18.0
gunicorn==23.0.0
hypothesis==6.123.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Levenshtein==0.26.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
